+++ NeuralNetwork.h
@@ -13,7 +13,7 @@ private:
 	int OutputLength;
 
 public:
-	NeuralNetwork(long shapeLength, long shape[], bool deleteShapeArr, ActivationFunctions::ActivationFunction activationFunction, double bias, double minWeight, double weightClosestTo0, double maxWeight)
+	NeuralNetwork(long shapeLength, long shape[], bool deleteShapeArr, ActivationFunctions::ActivationFunction activationFunction, float bias, float minWeight, float weightClosestTo0, float maxWeight)
 	{
 		Neurons = list<list<Neuron>>();
 		for (long i = 1; i < shapeLength; i++)
@@ -42,13 +42,13 @@ private:
 	}
 
 public:
-	double* Execute(double* input)
+	float* Execute(float* input)
 	{
-		tuple<double**, double**> storedExecution = ExecuteStore(input);
-		double** linears = get<0>(storedExecution);
-		double** activations = get<1>(storedExecution);
+		tuple<float**, float**> storedExecution = ExecuteStore(input);
+		float** linears = get<0>(storedExecution);
+		float** activations = get<1>(storedExecution);
 
-		double* output = activations[GetNetworkLayerCount()];
+		float* output = activations[GetNetworkLayerCount()];
 
 		delete[] linears[0];
 		for (size_t i = 1; i < GetNetworkLayerCount(); i++)
@@ -68,20 +68,20 @@ public:
 	/// </summary>
 	/// <param name="input"></param>
 	/// <returns>tuple(networkLinears, networkActivations)</returns>
-	tuple<double**, double**> ExecuteStore(double* input)
+	tuple<float**, float**> ExecuteStore(float* input)
 	{
-		double** networkLinearFunctions = new double* [GetNetworkLayerCount()];
-		double** networkActivations = new double* [GetNetworkLayerCount() + 1];
+		float** networkLinearFunctions = new float* [GetNetworkLayerCount()];
+		float** networkActivations = new float* [GetNetworkLayerCount() + 1];
 		networkActivations[0] = input;
 		auto layerIter = Neurons.begin();
 		for (long i = 0; i < GetNetworkLayerCount(); i++, layerIter++)
 		{
-			tuple<double*, double*> layerExecutionResults = ExecuteStoreLayer(layerIter, networkActivations);
+			tuple<float*, float*> layerExecutionResults = ExecuteStoreLayer(layerIter, networkActivations);
 			networkLinearFunctions[i] = get<0>(layerExecutionResults);
 			networkActivations[i + 1] = get<1>(layerExecutionResults);
 		}
 
-		tuple<double**, double**> output(networkLinearFunctions, networkActivations);
+		tuple<float**, float**> output(networkLinearFunctions, networkActivations);
 		return output;
 	}
 
@@ -91,32 +91,32 @@ public:
 	/// <param name="layerIter"></param>
 	/// <param name="networkActivations"></param>
 	/// <returns>tuple(linearFunctions, networkActivations)</returns>
-	tuple<double*, double*> ExecuteStoreLayer(list<list<Neuron>>::iterator layerIter, double** networkActivations)
+	tuple<float*, float*> ExecuteStoreLayer(list<list<Neuron>>::iterator layerIter, float** networkActivations)
 	{
 		list<Neuron> layer = (*layerIter);
 		long layerLength = layer.size();
-		double* layerLinears = new double[layerLength];
-		double* layerActivations = new double[layerLength];
+		float* layerLinears = new float[layerLength];
+		float* layerActivations = new float[layerLength];
 
 		auto neuronIter = layer.begin();
 		for (long i = 0; neuronIter != layer.end(); i++, neuronIter++)
 		{
-			tuple<double, double> neuronExecutionResults = (*neuronIter).ExecuteStore(networkActivations, ActivationFunction);
+			tuple<float, float> neuronExecutionResults = (*neuronIter).ExecuteStore(networkActivations, ActivationFunction);
 			layerLinears[i] = get<0>(neuronExecutionResults);
 			layerActivations[i] = get<1>(neuronExecutionResults);
 		}
-		tuple<double*, double*> layerExecutionResults(layerLinears, layerActivations);
+		tuple<float*, float*> layerExecutionResults(layerLinears, layerActivations);
 		return layerExecutionResults;
 	}
 
-	NeuralNetwork GetGradients(double* X, double* Y, Cost::CostFunction costFunction)
+	NeuralNetwork GetGradients(float* X, float* Y, Cost::CostFunction costFunction)
 	{
-		tuple<double**, double**> executionResults = ExecuteStore(X);
-		double** networkLinears, **networkActivations;
+		tuple<float**, float**> executionResults = ExecuteStore(X);
+		float** networkLinears, **networkActivations;
 		networkLinears = get<0>(executionResults);
 		networkActivations = get<1>(executionResults);
 
-		double* cost = Derivatives::DerivativeOf(GetNetworkOutputLength(), get<1>(executionResults)[GetNetworkLayerCount()], Y, costFunction);
+		float* cost = Derivatives::DerivativeOf(GetNetworkOutputLength(), get<1>(executionResults)[GetNetworkLayerCount()], Y, costFunction);
 		NeuralNetwork gradients = GetGradients(networkLinears, networkActivations, cost);
 
 		delete[] networkLinears[0];
@@ -133,15 +133,15 @@ public:
 		return gradients;
 	}
 
-	NeuralNetwork GetGradients(double** networkLinears, double** networkActivations, double* costGradients)
+	NeuralNetwork GetGradients(float** networkLinears, float** networkActivations, float* costGradients)
 	{
 		auto layerIterator = Neurons.begin();
-		double** neuronCosts = new double* [GetNetworkLayerCount() + 1];
+		float** neuronCosts = new float* [GetNetworkLayerCount() + 1];
 
 		neuronCosts[GetNetworkLayerCount()] = costGradients;
 		for (long i = 0; i < GetNetworkLayerCount(); i++)
 		{
-			neuronCosts[i] = new double[(*layerIterator).size()];
+			neuronCosts[i] = new float[(*layerIterator).size()];
 		}
 
 		long layerI = GetNetworkLayerCount();
@@ -169,7 +169,7 @@ public:
 	/// <param name="networkLinears"></param>
 	/// <param name="networkActivations"></param>
 	/// <returns></returns>
-	list<Neuron> CalculateLayerGradients(long layerI, list<list<Neuron>>::iterator layerIterator, double** neuronsGradient, double** networkLinears, double** networkActivations)
+	list<Neuron> CalculateLayerGradients(long layerI, list<list<Neuron>>::iterator layerIterator, float** neuronsGradient, float** networkLinears, float** networkActivations)
 	{
 		list<Neuron> layer = (*layerIterator);
 		list<Neuron> gradientsLayer = list<Neuron>();
@@ -178,11 +178,11 @@ public:
 		long i = 0;
 		while (neuronIterator != layer.end())
 		{
-			double linearFunction = networkLinears[layerI - 1][i];
-			double activation = neuronsGradient[layerI][i];
+			float linearFunction = networkLinears[layerI - 1][i];
+			float activation = neuronsGradient[layerI][i];
 
 			Neuron neuron = (*neuronIterator);
-			tuple<double, list<double>, list<double>> gradients = neuron.GetGradients(networkActivations, linearFunction, activation, ActivationFunction);
+			tuple<float, list<float>, list<float>> gradients = neuron.GetGradients(networkActivations, linearFunction, activation, ActivationFunction);
 
 			Neuron gradientNeuron = Neuron(get<0>(gradients), list<long>(), list<long>(), get<1>(gradients));
 
@@ -203,7 +203,7 @@ public:
 	}
 
 public:
-	void ApplyGradients(NeuralNetwork gradients, double learningRate)
+	void ApplyGradients(NeuralNetwork gradients, float learningRate)
 	{
 		auto layerIterator = Neurons.begin();
 		auto gLayerIterator = gradients.Neurons.begin();
